调整代码

+++ demo.Core/Dto/UserAddOutPutDto.go
@@ -3,5 +3,5 @@ package dto
 type UserAddOutPutDto struct {
 	Success bool   `json:"success"` //姓名
 	Error   string `json:"error"`   //性别
-	Id      int    `json:"userid"`  //年龄
+	Id      int64  `json:"userid"`  //年龄
 }
